Make TLS Endpoint generic to TLS provider
